+++ libs/video-player-core/src/utils/index.ts
@@ -3,5 +3,6 @@
  */
 
 export * from './format-time'
+export * from './frame-step'
 export * from './media-url'
 export * from './srt-to-vtt'